refactor: extract common bundle server utilities
- Create src/utils/bundle-server.ts with shared functions:
  - findBundles(): discover bundle directories
  - getMimeType(): get MIME type from file extension
  - isValidFile(): check if path is a valid file
  - createFileStream(): create read stream for file
- Update preview.ts and vite.config.ts to use shared utilities
- Reduces code duplication between dev and production servers
